# [perf] remove extra combine logic