@lifeart 2023-12-05
@chancancode could we have only one change from this PR?
eager return if item already in set: `this.tags.has(tag)) return`