@chancancode 2023-12-05
No that's what I was saying in the second half of the response, I specifically don't think we need to/should be doing that.

Unless we have evidence that this ends up being a hot spot in macro app benchmarks, I don't think that kind of changes are a good idea to make based on just those microbenchmarks you did. Even if you set up the benchmark perfectly it is still too small to be meaningful, and even if that happens to be slightly faster in the version of the browser you tested, that can easily change in a different browser, or in a different version. The current code is making a single call to `Set.prototype.add`, maybe it is somehow not particularly optimized, but if you switch it to make two separate calls in the happy path, that could easily end up being slower a few versions later when `.add` is optimized better, and no one has time to keep track of these things that are baked deeply in to the codebase.

So we just do the straight forward thing everywhere unless there is specific evidence that one spot needs to be optimized in a particular way, and even then you should heavily document why that deviates from the normal pattern in code comments, etc.