@NullVoxPopuli 2023-12-05
I'm re-opening because there is no evidence one way or another, and the opening of this PR is to communicate that we need to run benchmarks on this, before and after.