@lifeart 2023-12-05
Some extra notes here, I found `V8` implementation details for `has` and `add`:

https://github.com/nodejs/node/blob/1e31a01f898a3586faf01480928161f2f5ef965d/deps/v8/src/builtins/builtins-collections-gen.cc#L1730

### HAS
<img width="523" alt="image" src="https://github.com/glimmerjs/glimmer-vm/assets/1360552/d12d9db0-1d54-4687-bd54-ccea96dd009b">

### ADD
![telegram-cloud-photo-size-2-5438145921155191680-y](https://github.com/glimmerjs/glimmer-vm/assets/1360552/0bba2dde-5dd2-4ecf-9a25-775bb46f6a70)

### ADD_TO_ORDERED_HASH_TABLE
<img width="598" alt="image" src="https://github.com/glimmerjs/glimmer-vm/assets/1360552/019d1842-91b1-4073-852e-fa9a2846039e">


And it seems `has` is really faster in scope of logic. But I agree with @chancancode to not move here util we verify it really make difference in macrobench.