@wycats 2023-12-05
A few things:

- At minimum, we should wait until the errors feature has landed before considering this kind of change, because the errors feature touches on a lot of the core infrastructure.
- I agree with Godfrey that there should be a high bar for tweaking the deep tags/reference implementation. Changes should either come from feature needs or after identifying performance issues at a macro level and pinning them on a piece of low-level implementation.
- At bare minimum, changes like this should come along with reproducible benchmarks showing that the change makes a difference. Performance improvements also need to be verified in terms of higher-level benchmarks.

@lifeart your analysis of the v8 codebase is impressive and a really good start for the work that would go into a possible optimization.

Let's circle back around once the errors feature has landed and figure out what it would take to verify that the performance benefits are significant enough to meet the "changes to low-level features" bar.

In the meantime, you can feel free to apply the change to a local version of Ember and see if it impacts any of the ecosystem's macro-benchmarks (such as the https://github.com/krausest/js-framework-benchmark). Any information along those lines would be a good head start.