@lifeart 2023-12-23
### Bench results (23 Dec 2023)

```
    Benchmark Results Summary    

duration phase no difference [-66ms to 37ms]
render1000Items1End phase no difference [-7ms to 6ms]
clearItems1End phase no difference [0ms to 1ms]
render1000Items2End phase no difference [-4ms to 6ms]
clearItems2End phase no difference [-1ms to 0ms]
render10000Items1End phase estimated improvement -32ms [-65ms to -3ms] OR -0.83% [-1.69% to -0.08%]
clearItems3End phase no difference [-1ms to 2ms]
render1000Items3End phase estimated regression +9ms [2ms to 17ms] OR +2.03% [0.46% to 3.92%]
append1000Items1End phase no difference [-3ms to 9ms]
updateEvery10thItem1End phase no difference [0ms to 1ms]
selectFirstRow1End phase estimated regression +7ms [5ms to 8ms] OR +16.92% [12.61% to 21.22%]
selectSecondRow1End phase estimated improvement -4ms [-5ms to -4ms] OR -32.48% [-38.47% to -26.39%]
removeFirstRow1End phase no difference [-5ms to 0ms]
removeSecondRow1End phase no difference [0ms to 2ms]
swapRows1End phase no difference [-2ms to 1ms]
clearItems4End phase estimated regression +1ms [0ms to 2ms] OR +2.38% [0.41% to 4.29%]
paint phase no difference [0ms to 0ms]
```

[artifact-1.pdf](https://github.com/glimmerjs/glimmer-vm/files/13759228/artifact-1.pdf)